# api/swagger.yaml
openapi: 3.0.3
info:
  title: GophKeeper API
  description: |
    REST API менеджера паролей GophKeeper.
    Полезная нагрузка записей сейфа передаётся в зашифрованном виде (client-side encryption).
    Дополнительно доступен бинарный протокол синхронизации (encoding/gob) на `/api/v1/sync/binary`.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local server
paths:
  /health:
    get:
      summary: Health check
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
  /api/v1/register:
    post:
      summary: Register user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Credentials"
      responses:
        "200":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "409":
          description: Login already taken
  /api/v1/login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Credentials"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          description: Invalid credentials
  /api/v1/items:
    get:
      summary: List all vault items
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VaultItem"
    post:
      summary: Upsert vault item
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VaultItem"
      responses:
        "200":
          description: Saved
  /api/v1/items/{id}:
    get:
      summary: Get vault item by id
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultItem"
        "404":
          description: Not found
  /api/v1/sync:
    post:
      summary: Bidirectional JSON sync
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncRequest"
      responses:
        "200":
          description: Sync result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncResponse"
  /api/v1/sync/binary:
    post:
      summary: Bidirectional binary (gob) sync
      description: Тело запроса и ответа - gob-сериализация SyncRequest/SyncResponse.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: Sync result
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Credentials:
      type: object
      required: [login, password]
      properties:
        login:
          type: string
        password:
          type: string
          format: password
    TokenResponse:
      type: object
      properties:
        token:
          type: string
    VaultItem:
      type: object
      required: [id, version, payload]
      properties:
        id:
          type: string
          format: uuid
        version:
          type: integer
          format: int64
        updated_at:
          type: string
          format: date-time
        deleted:
          type: boolean
        payload:
          type: string
          format: byte
          description: AES-GCM encrypted ItemPayload
    SyncRequest:
      type: object
      properties:
        since:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/VaultItem"
    SyncResponse:
      type: object
      properties:
        server_time:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/VaultItem"